Modify selection validation logic
Replace `checkSeatSelectionValidity`, which checks the effect of an
operation, that can potentially invalidate the selection, before
applying that operation.

It is replaced with `auditSeatSelection`, which allows the operation
to be applied, but the modified states are then checked before
sending updates to the user.

This change has a few benefits: all selection validation logic are
now encapsulated within `auditSeatSelection`, and we can now detect
warnings (previously `SeatSelectionResult`) which are not always
caused by a new selection (e.g. a seat may become isolated after
a call to `updateTakenStatus`).