Create first draft of seating plan component (#22)
* Add BaseSeatingPlanManager and related types

* Move components directory

* Add draft of SeatingPlan component

* Modify SeatingPlan to use Context

* Add draft of CustomerSeatingPlanInterface

* Modify some types

* Create mock types and page

* Edit mock styles and fix size issues

* Fix components failing to re-render upon selection update

This solution is a bit hacky, will probably change

* Replace dummy map image

* Fix typo in image source

* Change text color for visibility

* Edit types and update seat-related fields

This commit changes a few things to reflect new information about
business logic specifics since the date of the last commit:
- Replace BaseSeatState with SeatState and is no longer inherited
- Require some components to specify the BaseSeatingPlanManager
  subtype
- Add level as a seat metadata field
- Add more fields and methods into CustomerSeatingPlanManager
- Relocate some props and component codes

* Edit mock types and develop mock page

Relocate some props for more modularity

* Add levels data and refactoring

* Add label to seat, level, and type

Also fixed a typo

* Simplify display of seat selection

* Change seat type to category

* Modify selection validation logic

Replace `checkSeatSelectionValidity`, which checks the effect of an
operation, that can potentially invalidate the selection, before
applying that operation.

It is replaced with `auditSeatSelection`, which allows the operation
to be applied, but the modified states are then checked before
sending updates to the user.

This change has a few benefits: all selection validation logic are
now encapsulated within `auditSeatSelection`, and we can now detect
warnings (previously `SeatSelectionResult`) which are not always
caused by a new selection (e.g. a seat may become isolated after
a call to `updateTakenStatus`).

* Implement minimal demo