Implement seat selection page (#48)
* Copy mock-seating-plan to ticket/select page

* Edit type names for seating plan

Differentiate clearly between types defined in `lib/` and in
`components/seating-plan/` by prefixing `UI` to the type name.

* Cache isAvailable values and getSeatsMetadata

* Create a new collection "caches" with a document "seats.isAvailable"
  that stores the values of isAvailable for each seat ID
* Make the endpoint `getSeatsMetadata` permanently cached
* Fix a bug where `leftId` and `rightId` are undefined instead of null

* Connect select page to database

* Handle seat selection validation and implement notifications

* Add comment in email API

* Fix typo in updating isAvailable cache

* Add logic to block warnings when awaiting server confirmation

This is needed because currently if the database successfully
reserves the seats, the change in their availability is detected
by the client often before the response is received, and this
change prompts the SEAT_TAKEN warning, which is not what we want
the user to see.

* Add temporary print statement